requirements.txt
numpy==1.24.4
opencv_contrib_python==4.9.0.80
Pillow==9.1.0
Pillow==10.4.0
Requests==2.32.3
streamlit==1.38.0
